Add files via upload
This is a desktop application written in C++ that lets you add, subtract, multiply or divide the elements of 2 equally long vectors regardless how long they are. It also is programmed to recognize invalid inputs by the user and displays an erorr as well as enabling another correct input chance.